Introduces nb_dofs and nb_force_sensors config parameters.
To be a bit flexible (for use with Romeo) the binding with the SoT
is based upon parameters specificied in rtc.conf.
More specifically this commit introduces two parameters:
* robot.nb_dofs: number of dofs of the robot.
* robot.nb_forces_sensors: number of force sensors.
The latter one is not yet fully flexible.
The number of input ports should be automatically based
upon this value, but currently it is hardcoded to 4.

This current version do not segfault with libsot-hrp2-14-controller.so
but freezes.