Correct the number of dofs for the hrp2-10/14.
The number of dofs correspond to the real number of dof for the robot.
ie, it is equal to the number of dof of the small model + 10, not
the number of dofs of the small model + 6 (= num of waist dofs).

This is based on the size of the reference vector given by
rtprint /localhost/kpg.rtc:qRef
when running the default sample script.

Besides, using the former number of dofs, the stabilisator may
not work properly.